+++ src/User/RegisterUserCommand.php
@@ -43,6 +43,9 @@ final class RegisterUserCommand
     /** @var string */
     private $externalIdentifier;
 
+    /** @var bool|null */
+    private $isProfessional;
+
     /** @return string|null */
     public function getPhone(): ?string
     {
@@ -229,6 +232,18 @@ final class RegisterUserCommand
     public function setExternalIdentifier(?string $externalIdentifier): self
     {
         $this->externalIdentifier = $externalIdentifier;
+
+        return $this;
+    }
+
+    public function getIsProfessional(): ?bool
+    {
+        return $this->isProfessional;
+    }
+
+    public function setIsProfessional(?bool $isProfessional): self
+    {
+        $this->isProfessional = $isProfessional;
 
         return $this;
     }